Separate pixelmon compat from main mixins
