+++ src/main/java/net/coderbot/iris/compat/pixelmon/mixin/OculusPixelmonCompatMixinPlugin.java
@@ -1,4 +1,4 @@
-package net.coderbot.iris.mixin.compat;
+package net.coderbot.iris.compat.pixelmon.mixin;
 
 import net.minecraftforge.fml.loading.FMLLoader;
 import org.objectweb.asm.tree.ClassNode;
@@ -8,7 +8,7 @@ import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
 import java.util.List;
 import java.util.Set;
 
-public class CompatMixinPlugin implements IMixinConfigPlugin {
+public class OculusPixelmonCompatMixinPlugin implements IMixinConfigPlugin {
     @Override
     public void onLoad(String mixinPackage) {
 
@@ -21,10 +21,7 @@ public class CompatMixinPlugin implements IMixinConfigPlugin {
 
     @Override
     public boolean shouldApplyMixin(String targetClassName, String mixinClassName) {
-        int startIndex = mixinClassName.indexOf("compat.") + "compat.".length();
-        int endIndex = mixinClassName.indexOf(".", startIndex);
-        String modid = mixinClassName.substring(startIndex, endIndex);
-        return FMLLoader.getLoadingModList().getModFileById(modid) != null;
+        return FMLLoader.getLoadingModList().getModFileById("pixelmon") != null;
     }
 
     @Override
