--- src/main/resources/mixins.oculus.compat.json
@@ -1,14 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "net.coderbot.iris.mixin.compat",
-  "plugin": "net.coderbot.iris.mixin.compat.CompatMixinPlugin",
-  "refmap": "oculus-refmap.json",
-  "compatibilityLevel": "JAVA_8",
-  "client": [
-    "pixelmon.MixinNormalizedFace"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
